Rearrange class into separate modules
